clean-up pages and general formatting
