--- src/components/CleanUIComponents/Donut/style.module.scss
@@ -1,34 +1,0 @@
-@import 'assets/styles/mixins.scss';
-
-.donut {
-  display: inline-block;
-  width: rem(16);
-  height: rem(16);
-  border-radius: 100%;
-  border: 4px solid $default;
-  position: relative;
-  top: rem(2);
-  margin-right: rem(4);
-}
-
-.default {
-  border-color: $default;
-}
-.primary {
-  border-color: $primary;
-}
-.secondary {
-  border-color: $secondary;
-}
-.success {
-  border-color: $success;
-}
-.warning {
-  border-color: $warning;
-}
-.danger {
-  border-color: $danger;
-}
-.info {
-  border-color: $info;
-}
