+++ src/components/CleanUIComponents/InfoCard2/data.json
@@ -8,7 +8,7 @@
     "text": "Lorem ipsum dolor sit amet, consectetur adipisicing elit. Asperiores, vero."
   },
   "commerceStats": {
-    "title": "LIFETIME REVENUE",
-    "count": "$366,942.00"
+    "title": "ANNUAL REVENUE",
+    "count": "$66,323.00"
   }
 }
