--- src/components/CleanUIComponents/UserCard/data.json
@@ -1,7 +1,0 @@
-{
-  "user": {
-    "name": "Artour Junior Aniston",
-    "post": "Project Manager",
-    "avatar": "resources/images/avatars/1.jpg"
-  }
-}
