--- src/pages/ecommerce/product-details/data.json
@@ -1,86 +1,0 @@
-{
-  "name": "We market your suite!",
-  "rate": 5,
-  "images": [ "https://a0.muscache.com/im/pictures/b893adfd-e749-465b-851d-23db01ad8cc2.jpg?aki_policy=x_large",
-              "https://a0.muscache.com/im/pictures/f3e7f609-0be1-4a7e-a81b-2572d3ebe724.jpg?aki_policy=large",
-              "https://a0.muscache.com/im/pictures/4199850b-2294-4949-964f-092b45f85bf4.jpg?aki_policy=x_large",
-              "https://a0.muscache.com/im/pictures/078aa2c0-d1c1-4d2f-99b3-c058743801bf.jpg?aki_policy=x_large"],
-  "price": "189.95/night",
-  "shortDescr": "We provide marketing, customized posting with pricing based on data-driven strategies, inventory, house-keeping, maintenance, guest check-in service, guest management and more.",
-  "services": [
-    {
-    "name": "* ",
-    "value": "Short-term vacation rental program"
-    },
-    {
-    "name": "* ",
-    "value": "Long-term off-season rentals (Sept 1st -April 30th)"
-    },
-    {
-    "name": "* ",
-    "value": "Marketing of your property on third party websites"
-    },
-    {
-    "name": "* ",
-    "value": "Housekeeping services"
-    },
-    {
-    "name": "* ",
-    "value": "Property checks"
-    },
-    {
-    "name": "* ",
-    "value": "Maintenance"
-    },
-    {
-    "name": "* ",
-    "value": "Tax collection and remittance"
-    },
-    {
-    "name": "* ",
-    "value": "Linen Service"
-    },
-    {
-      "name": "* ",
-      "value": "Check in service"
-    },
-    {
-      "name": "* ",
-      "value": "Guest interaction"
-    },
-    {
-      "name": "* ",
-      "value": "Monthly and annual statements to owners"
-    }
-  ],
-  "properties": [
-    {
-      "name": "SuperHost Rating",
-      "value": "262 of 275 reviews are 5 star"
-    },
-    {
-      "name":"Feature 2",
-      "value": "something"
-    },
-    {
-      "name":"Feature 3",
-      "value": "something"
-    },
-    {
-      "name":"Feature 3",
-      "value": "something"
-    },
-    {
-      "name":"Feature 4",
-      "value": "something"
-    },
-    {
-      "name":"Feature 5",
-      "value": "something"
-    },
-    {
-      "name":"Feature 6",
-      "value": "something"
-    }
-  ]
-}
